some fun new stuff!!

+++ Interstation-Two-WW2/code/game/gamemodes/WW2/ww2.dm
@@ -292,7 +292,8 @@
 		// no tanks on lowpop
 		if (!istype(aspect, /datum/game_aspect/ww2/no_tanks))
 			if (clients.len <= TANK_LOWPOP_THRESHOLD)
-				for (var/obj/tank/T in world)
-					qdel(T)
-				world << "<i>Due to lowpop, there are no tanks.</i>"
+				if (locate(/obj/tank) in world)
+					for (var/obj/tank/T in world)
+						qdel(T)
+					world << "<i>Due to lowpop, there are no tanks.</i>"
